move op tests to test_blocks
